frontend: move utils simplemarkup to markup
Move utils/simplemarkup.ts -> utils/markup.ts to keep the history.

Goal is to add another export for multilinemarkup.
