+++ frontends/web/src/components/alert/Alert.tsx
@@ -17,7 +17,7 @@
 
 import { Component, h, RenderableProps } from 'preact';
 import { translate, TranslateProps } from '../../decorators/translate';
-import { SimpleMarkup } from '../../utils/simplemarkup';
+import { SimpleMarkup } from '../../utils/markup';
 import { Dialog } from '../dialog/dialog';
 import * as style from '../dialog/dialog.module.css';
 import { Button } from '../forms';
