+++ frontends/web/src/routes/device/bitbox01/setup/security-information.tsx
@@ -21,7 +21,7 @@ import { LanguageSwitch } from '../../../../components/language/language';
 import { Header } from '../../../../components/layout';
 import { Message } from '../../../../components/message/message';
 import { translate,  TranslateProps } from '../../../../decorators/translate';
-import { SimpleMarkup } from '../../../../utils/simplemarkup';
+import { SimpleMarkup } from '../../../../utils/markup';
 import * as style from '../bitbox01.module.css';
 
 interface SecurityInformationProps {
